Move login and signup pages to new "auth" feature folder
