+++ frontend/src/App.tsx
@@ -17,9 +17,9 @@ import PlayBot from "./pages/protected/PlayBot.tsx";
 import GameHistory from "./pages/protected/GameHistory/GameHistory.tsx";
 import ViewGame from "./pages/protected/ViewGame.tsx";
 import ChallengeWebsocketProvider from "./components/wrappers/ChallengeWebsocketProvider.tsx";
-import Login from "./pages/Login.tsx";
+import Login from "./features/auth/pages/Login.tsx";
 import GuestHomePage from "./pages/GuestHomePage/GuestHomePage.tsx";
-import Signup from "./pages/Signup.tsx";
+import Signup from "./features/auth/pages/Signup.tsx";
 import Dashboard from "./pages/protected/Dashboard/Dashboard.tsx";
 import GameSetup from "./pages/protected/GameSetup.tsx";
 import NotFound from "./pages/NotFound.tsx";
