+++ frontend/src/features/auth/pages/Login.tsx
@@ -1,0 +1,7 @@
+import AuthForm from "../../forms/AuthForm";
+
+function Login() {
+	return <AuthForm method="Login" />;
+}
+
+export default Login;
