+++ frontend/src/features/auth/pages/Signup.tsx
@@ -1,0 +1,7 @@
+import AuthForm from "../../forms/AuthForm";
+
+function Signup() {
+	return <AuthForm method="Signup" />;
+}
+
+export default Signup;
