--- frontend/src/pages/Login.tsx
@@ -1,7 +1,0 @@
-import AuthForm from "../features/forms/AuthForm";
-
-function Login() {
-	return <AuthForm method="Login" />;
-}
-
-export default Login;
