--- frontend/src/pages/Signup.tsx
@@ -1,7 +1,0 @@
-import AuthForm from "../features/forms/AuthForm";
-
-function Signup() {
-	return <AuthForm method="Signup" />;
-}
-
-export default Signup;
